+++ src/main/kotlin/blackjack/domain/model/participant/Player.kt
@@ -1,7 +1,10 @@
 package blackjack.domain.model.participant
 
 import blackjack.domain.model.card.Card
-import blackjack.domain.model.progress.BetAmount
+import blackjack.domain.model.card.CardStatus
+import blackjack.domain.model.participant.bet.BetAmount
+import blackjack.domain.model.participant.bet.Profit
+import blackjack.domain.model.participant.bet.ProfitRate
 
 class Player(
     playerInfo: ParticipantInfo = ParticipantInfo(DEFAULT_NAME, BetAmount()),
