+++ src/main/kotlin/blackjack/view/Translator.kt
@@ -1,7 +1,6 @@
 package blackjack.view
 
 import blackjack.domain.model.card.Suit
-import blackjack.domain.model.progress.WinLoss
 import java.util.Locale
 
 object Translator {
@@ -19,20 +18,5 @@ object Translator {
                 }
 
             else -> suit.name
-        }
-
-    fun winLossLocalize(
-        winLoss: WinLoss,
-        locale: Locale,
-    ): String =
-        when (locale) {
-            Locale.KOREAN ->
-                when (winLoss) {
-                    WinLoss.WIN -> "승"
-                    WinLoss.DRAW -> "무"
-                    WinLoss.LOSE -> "패"
-                }
-
-            else -> winLoss.name
         }
 }
