+++ src/test/kotlin/blackjack/domain/model/progress/BetAmountTest.kt
@@ -1,5 +1,6 @@
 package blackjack.domain.model.progress
 
+import blackjack.domain.model.participant.bet.BetAmount
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.assertThrows
 import org.junit.jupiter.params.ParameterizedTest
