+++ src/test/kotlin/blackjack/domain/model/progress/BetHistoryTest.kt
@@ -1,19 +1,15 @@
 package blackjack.domain.model.progress
 
-import blackjack.domain.model.participant.Player
-import org.assertj.core.api.Assertions.assertThat
-import org.junit.jupiter.api.Test
-
 class BetHistoryTest {
-    @Test
-    fun `베팅 기록들을 추가하고 읽을 수 있다`() {
-        val betHistory = BetHistory()
-        val meda = Player("메다")
-        val cream = Player("크림")
-
-        betHistory.addBetLog(meda, BetAmount(1000))
-        betHistory.addBetLog(cream, BetAmount(2000))
-
-        assertThat(betHistory.playerBets).isEqualTo(mapOf(meda to 1000, cream to 2000))
-    }
+//    @Test
+//    fun `베팅 기록들을 추가하고 읽을 수 있다`() {
+//        val betHistory = BetHistory()
+//        val meda = Player("메다")
+//        val cream = Player("크림")
+//
+//        betHistory.addBetLog(meda, BetAmount(1000))
+//        betHistory.addBetLog(cream, BetAmount(2000))
+//
+//        assertThat(betHistory.playerBets).isEqualTo(mapOf(meda to 1000, cream to 2000))
+//    }
 }
